refactor(webgl): split webgl2-rendering-context.ts (1245 → 866 + 4 method-group modules)
Second slice of the long-running webgl2-rendering-context.ts split, sibling
to PR #273. Extracts the four method groups PR #273 explicitly left as
"future PR" candidates, using the same install*Methods(proto) pattern as
object-lifecycle.ts.

| Group | New file | Methods | LoC |
|---|---|---|---:|
| 3D texture upload | webgl2-context/tex-image-3d.ts | texImage3D, texSubImage3D, compressedTexImage3D, compressedTexSubImage3D, copyTexSubImage3D, texStorage2D, texStorage3D | 82 |
| Unsigned-int + non-square matrix uniforms | webgl2-context/uniform.ts | uniform1ui/2ui/3ui/4ui, uniform1uiv/2uiv/3uiv/4uiv, uniformMatrix2x3fv/3x2fv/2x4fv/4x2fv/3x4fv/4x3fv | 128 |
| clearBuffer family | webgl2-context/clear-buffer.ts | clearBufferfv, clearBufferiv, clearBufferuiv, clearBufferfi | 123 |
| Instanced-draw + framebuffer ops | webgl2-context/instanced-and-framebuffer-ops.ts | drawArraysInstanced, drawElementsInstanced, vertexAttribDivisor, vertexAttribIPointer, drawBuffers, drawRangeElements, blitFramebuffer, framebufferTextureLayer, invalidateFramebuffer, invalidateSubFramebuffer, readBuffer, renderbufferStorageMultisample | 214 |

37 methods totalling ~380 LoC moved out of the base file. Each new module
declaration-merges its <Group>Methods interface into WebGL2RenderingContext
via `declare module '../webgl2-rendering-context.js'` and exports an
install<Group>Methods(proto) function.

Base file webgl2-rendering-context.ts: 1245 → 866 LoC. Calls all five
install functions (the four new ones plus the installObjectLifecycleMethods
PR #273 added) after the class declaration. Each install function is
imported twice — once as a named import (for the call site) and once as a
side-effect import — so tsc preserves the augmentation in the emitted
.d.ts (same technique as PR #273 / the canvas2d-core split in PR #262).

File-local helpers (`OptionalClearBufferNative` interface, `DebugCounters`
interface, `debugGlEnabled()` function) moved verbatim into the module
that uses them. No `private` field had to be demoted because the only
private member touched by the extracted methods is referenced only by
methods that stayed in the base.

860/860 GJS webgl conformance tests still green.
