+++ packages/framework/webgl/src/ts/webgl2-rendering-context.ts
@@ -20,29 +20,6 @@ import { WebGLFramebuffer } from './webgl-framebuffer.js';
 import { Uint8ArrayToVariant, arrayToUint8Array, vertexCount, convertPixels, extractImageData, checkObject, premultiplyAlpha } from './utils.js';
 import { warnNotImplemented } from '@gjsify/utils';
 
-/** Internal debug counters attached to the context instance — opt-in via `globalThis.__GJSIFY_DEBUG_GL = true`. */
-interface DebugCounters {
-    __drawInstCount?: number;
-    __blitCount?: number;
-}
-
-/**
- * `clearBuffer{fv,iv,uiv,fi}` only ship in newer Vala bindings of `Gwebgl.WebGL2RenderingContext`.
- * The static GIR types may not include them yet; we feature-detect at runtime, so accept
- * "either present or absent" without falling back to `as any`.
- */
-interface OptionalClearBufferNative {
-    clearBufferfv?(buffer: GLenum, drawbuffer: GLint, values: number[]): void;
-    clearBufferiv?(buffer: GLenum, drawbuffer: GLint, values: number[]): void;
-    clearBufferuiv?(buffer: GLenum, drawbuffer: GLint, values: number[]): void;
-    clearBufferfi?(buffer: GLenum, drawbuffer: GLint, depth: number, stencil: number): void;
-}
-
-/** Read the global debug flag without leaking `any` into call sites. */
-function debugGlEnabled(): boolean {
-    return (globalThis as { __GJSIFY_DEBUG_GL?: boolean }).__GJSIFY_DEBUG_GL === true;
-}
-
 export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2RenderingContext {
 
     _native2: Gwebgl.WebGL2RenderingContext;
@@ -446,52 +423,6 @@ export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2Re
         dst.set(data.subarray(0, dst.byteLength));
     }
 
-    // ─── 3D Textures ──────────────────────────────────────────────────────
-
-    texImage3D(target: GLenum, level: GLint, internalformat: GLint, width: GLsizei, height: GLsizei, depth: GLsizei, border: GLint, format: GLenum, type: GLenum, pixels: ArrayBufferView | null): void {
-        if (pixels === null) {
-            this._native2.texImage3DNull(target, level, internalformat, width, height, depth, border, format, type);
-        } else {
-            this._native2.texImage3D(target, level, internalformat, width, height, depth, border, format, type, Uint8ArrayToVariant(new Uint8Array(pixels.buffer, pixels.byteOffset, pixels.byteLength)));
-        }
-    }
-
-    texSubImage3D(target: GLenum, level: GLint, xoffset: GLint, yoffset: GLint, zoffset: GLint, width: GLsizei, height: GLsizei, depth: GLsizei, format: GLenum, type: GLenum, pixels: ArrayBufferView | null): void {
-        if (pixels === null) return;
-        this._native2.texSubImage3D(target, level, xoffset, yoffset, zoffset, width, height, depth, format, type, Uint8ArrayToVariant(new Uint8Array(pixels.buffer, pixels.byteOffset, pixels.byteLength)));
-    }
-
-    compressedTexImage3D(target: GLenum, level: GLint, internalformat: GLenum, width: GLsizei, height: GLsizei, depth: GLsizei, border: GLint, _imageSize: GLsizei, data: ArrayBufferView): void {
-        this._native2.compressedTexImage3D(target, level, internalformat, width, height, depth, border, Uint8ArrayToVariant(new Uint8Array(data.buffer, data.byteOffset, data.byteLength)));
-    }
-
-    compressedTexSubImage3D(target: GLenum, level: GLint, xoffset: GLint, yoffset: GLint, zoffset: GLint, width: GLsizei, height: GLsizei, depth: GLsizei, format: GLenum, _imageSize: GLsizei, data: ArrayBufferView): void {
-        this._native2.compressedTexSubImage3D(target, level, xoffset, yoffset, zoffset, width, height, depth, format, Uint8ArrayToVariant(new Uint8Array(data.buffer, data.byteOffset, data.byteLength)));
-    }
-
-    copyTexSubImage3D(target: GLenum, level: GLint, xoffset: GLint, yoffset: GLint, zoffset: GLint, x: GLint, y: GLint, width: GLsizei, height: GLsizei): void {
-        this._native2.copyTexSubImage3D(target, level, xoffset, yoffset, zoffset, x, y, width, height);
-    }
-
-    texStorage2D(target: GLenum, levels: GLsizei, internalformat: GLenum, width: GLsizei, height: GLsizei): void {
-        this._native2.texStorage2D(target, levels, internalformat, width, height);
-        // Update JS-side metadata so _updateFramebufferAttachments / _preCheckFramebufferStatus
-        // can see valid dimensions. Without this, w/h stay 0 and the attachment is cleared.
-        const texture = this._getTexImage(target);
-        if (texture) {
-            for (let lvl = 0; lvl < levels; lvl++) {
-                texture._levelWidth[lvl] = Math.max(1, width >> lvl);
-                texture._levelHeight[lvl] = Math.max(1, height >> lvl);
-            }
-            texture._format = this.RGBA; // base format; type varies but unused by our completeness check
-            texture._type = this.UNSIGNED_BYTE;
-        }
-    }
-
-    texStorage3D(target: GLenum, levels: GLsizei, internalformat: GLenum, width: GLsizei, height: GLsizei, depth: GLsizei): void {
-        this._native2.texStorage3D(target, levels, internalformat, width, height, depth);
-    }
-
     // ─── WebGL2 texImage2D / texSubImage2D overrides ─────────────────────
     // The base WebGL1 implementation rejects WebGL2 sized internal formats
     // (e.g. RGBA8, RGB8, SRGB8_ALPHA8) and requires format === internalFormat.
@@ -651,328 +582,6 @@ export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2Re
         }
 
         this._gl.texSubImage2D(target, level, xoffset, yoffset, width, height, format, type, Uint8ArrayToVariant(data));
-    }
-
-    framebufferTextureLayer(target: GLenum, attachment: GLenum, texture: WebGLTexture | null, level: GLint, layer: GLint): void {
-        this._native2.framebufferTextureLayer(target, attachment, texture ? (texture as unknown as WebGLTexture)._ : 0, level, layer);
-    }
-
-    // ─── Instancing & Advanced Draw ───────────────────────────────────────
-
-    drawArraysInstanced(mode: GLenum, first: GLint, count: GLsizei, instanceCount: GLsizei): void {
-        if (first < 0 || count < 0 || instanceCount < 0) { this.setError(this.INVALID_VALUE); return; }
-        if (!this._checkStencilState()) return;
-        const rc = vertexCount(this, mode, count);
-        if (rc < 0) { this.setError(this.INVALID_ENUM); return; }
-        if (!this._framebufferOk()) return;
-        if (count === 0 || instanceCount === 0) return;
-        if (!this._checkVertexAttribState((count + first - 1) >>> 0)) return;
-        if (debugGlEnabled()) {
-            const dbg = this as DebugCounters;
-            const n = dbg.__drawInstCount = (dbg.__drawInstCount | 0) + 1;
-            if (n <= 5 || n % 100 === 0) console.log(`[WebGL] drawArraysInstanced #${n} count=${rc} instances=${instanceCount} fbo=${this._activeFramebuffer?._ ?? '_gtkFbo'}`);
-        }
-        this._native2.drawArraysInstanced(mode, first, rc, instanceCount);
-    }
-
-    drawElementsInstanced(mode: GLenum, count: GLsizei, type: GLenum, offset: GLintptr, instanceCount: GLsizei): void {
-        if (count < 0 || offset < 0 || instanceCount < 0) { this.setError(this.INVALID_VALUE); return; }
-        if (!this._checkStencilState()) return;
-        const elementBuffer = this._vertexObjectState._elementArrayBufferBinding;
-        if (!elementBuffer) { this.setError(this.INVALID_OPERATION); return; }
-
-        let elementData: Uint8Array | Uint16Array | Uint32Array | null = null;
-        let adjustedOffset = offset;
-        if (type === this.UNSIGNED_SHORT) {
-            if (adjustedOffset % 2) { this.setError(this.INVALID_OPERATION); return; }
-            adjustedOffset >>= 1;
-            elementData = new Uint16Array(elementBuffer._elements.buffer);
-        } else if (type === this.UNSIGNED_INT) {
-            if (adjustedOffset % 4) { this.setError(this.INVALID_OPERATION); return; }
-            adjustedOffset >>= 2;
-            elementData = new Uint32Array(elementBuffer._elements.buffer);
-        } else if (type === this.UNSIGNED_BYTE) {
-            elementData = elementBuffer._elements;
-        } else {
-            this.setError(this.INVALID_ENUM);
-            return;
-        }
-
-        let reducedCount = count;
-        switch (mode) {
-            case this.TRIANGLES: if (count % 3) reducedCount -= (count % 3); break;
-            case this.LINES: if (count % 2) reducedCount -= (count % 2); break;
-            case this.POINTS: break;
-            case this.LINE_LOOP: case this.LINE_STRIP:
-                if (count < 2) { this.setError(this.INVALID_OPERATION); return; }
-                break;
-            case this.TRIANGLE_FAN: case this.TRIANGLE_STRIP:
-                if (count < 3) { this.setError(this.INVALID_OPERATION); return; }
-                break;
-            default:
-                this.setError(this.INVALID_ENUM);
-                return;
-        }
-
-        if (!this._framebufferOk()) return;
-        if (reducedCount === 0 || instanceCount === 0) { this._checkVertexAttribState(0); return; }
-        if ((reducedCount + adjustedOffset) >>> 0 > elementData.length) {
-            this.setError(this.INVALID_OPERATION);
-            return;
-        }
-
-        let maxIndex = 0;
-        for (let i = adjustedOffset; i < adjustedOffset + reducedCount; ++i) {
-            if (elementData[i] > maxIndex) maxIndex = elementData[i];
-        }
-        if (this._checkVertexAttribState(maxIndex)) {
-            this._native2.drawElementsInstanced(mode, reducedCount, type, offset, instanceCount);
-        }
-    }
-
-    vertexAttribDivisor(index: GLuint, divisor: GLuint): void {
-        this._native2.vertexAttribDivisor(index, divisor);
-    }
-
-    vertexAttribIPointer(index: GLuint, size: GLint, type: GLenum, stride: GLsizei, offset: GLintptr): void {
-        this._native2.vertexAttribIPointer(index, size, type, stride, offset);
-    }
-
-    drawBuffers(buffers: GLenum[]): void {
-        // GL_BACK (0x0405) is only valid for the window-system default framebuffer (FBO 0).
-        // GtkGLArea uses its own FBO (not FBO 0), so GL_BACK → GL_INVALID_OPERATION.
-        // Map GL_BACK → GL_COLOR_ATTACHMENT0 which is the attachment GTK's FBO uses.
-        let hasBack = false;
-        for (let i = 0; i < buffers.length; i++) {
-            if (buffers[i] === 0x0405) { hasBack = true; break; }
-        }
-        if (!hasBack) {
-            this._native2.drawBuffers(buffers);
-            return;
-        }
-        this._native2.drawBuffers(buffers.map(b => b === 0x0405 /* GL_BACK */ ? this.COLOR_ATTACHMENT0 : b));
-    }
-
-    drawRangeElements(mode: GLenum, start: GLuint, end: GLuint, count: GLsizei, type: GLenum, offset: GLintptr): void {
-        if (count < 0 || offset < 0) { this.setError(this.INVALID_VALUE); return; }
-        if (end < start) { this.setError(this.INVALID_VALUE); return; }
-        // Delegate to drawElements for full validation.
-        // drawRangeElements is just a hint to the driver about the index range.
-        this.drawElements(mode, count, type, offset);
-    }
-
-    blitFramebuffer(srcX0: GLint, srcY0: GLint, srcX1: GLint, srcY1: GLint, dstX0: GLint, dstY0: GLint, dstX1: GLint, dstY1: GLint, mask: GLbitfield, filter: GLenum): void {
-        if (debugGlEnabled()) {
-            // Check GL error before blit to isolate issues
-            const errBefore = this._gl.getError();
-            if (errBefore !== 0) console.log(`[WebGL] blitFramebuffer PRE-ERROR 0x${errBefore.toString(16)}`);
-        }
-        this._native2.blitFramebuffer(srcX0, srcY0, srcX1, srcY1, dstX0, dstY0, dstX1, dstY1, mask, filter);
-        if (debugGlEnabled()) {
-            const err = this._gl.getError();
-            const dbg = this as DebugCounters;
-            const n = dbg.__blitCount = (dbg.__blitCount | 0) + 1;
-            if (n <= 5) console.log(`[WebGL] blitFramebuffer #${n} src=(${srcX0},${srcY0},${srcX1},${srcY1}) readFbo=${this._activeReadFramebuffer?._ ?? '_gtkFbo'} err=${err === 0 ? 'OK' : '0x' + err.toString(16)}`);
-        }
-    }
-
-    // clearBuffer{fv,iv,uiv,fi} — WebGL2 methods for clearing specific
-    // framebuffer attachments. The native Vala binding does not expose the
-    // glClearBuffer* entry points yet, so we emulate the common cases via
-    // glClearColor/glClearDepth/glClearStencil + glClear. This is equivalent
-    // when the DRAW_FRAMEBUFFER has a single attachment per buffer type,
-    // which matches Excalibur's ExcaliburGraphicsContextWebGL.blitToScreen.
-    //
-    // Buffer target constants per WebGL2 spec (not on our class):
-    //   COLOR         = 0x1800
-    //   DEPTH         = 0x1801
-    //   STENCIL       = 0x1802
-    //   DEPTH_STENCIL = 0x84F9
-
-    clearBufferfv(buffer: GLenum, drawbuffer: GLint, values: Float32List, _srcOffset?: GLuint): void {
-        const n2 = this._native2 as Gwebgl.WebGL2RenderingContext & OptionalClearBufferNative;
-        if (typeof n2.clearBufferfv === 'function') {
-            n2.clearBufferfv(buffer, drawbuffer, Array.from(values) as number[]);
-            return;
-        }
-        const v = values as ArrayLike<number>;
-        if (buffer === 0x1800 /* COLOR */) {
-            const prev = this.getParameter(this.COLOR_CLEAR_VALUE) as Float32Array | number[] | null;
-            this.clearColor(v[0] ?? 0, v[1] ?? 0, v[2] ?? 0, v[3] ?? 0);
-            this.clear(this.COLOR_BUFFER_BIT);
-            if (prev) this.clearColor(prev[0], prev[1], prev[2], prev[3]);
-        } else if (buffer === 0x1801 /* DEPTH */) {
-            const prev = this.getParameter(this.DEPTH_CLEAR_VALUE) as number | null;
-            this.clearDepth(v[0] ?? 1);
-            this.clear(this.DEPTH_BUFFER_BIT);
-            if (prev !== null) this.clearDepth(prev);
-        }
-    }
-
-    clearBufferiv(buffer: GLenum, drawbuffer: GLint, values: Int32List, _srcOffset?: GLuint): void {
-        const n2 = this._native2 as Gwebgl.WebGL2RenderingContext & OptionalClearBufferNative;
-        if (typeof n2.clearBufferiv === 'function') {
-            n2.clearBufferiv(buffer, drawbuffer, Array.from(values) as number[]);
-            return;
-        }
-        if (buffer === 0x1802 /* STENCIL */) {
-            const v = values as ArrayLike<number>;
-            const prev = this.getParameter(this.STENCIL_CLEAR_VALUE) as number | null;
-            this.clearStencil(v[0] ?? 0);
-            this.clear(this.STENCIL_BUFFER_BIT);
-            if (prev !== null) this.clearStencil(prev);
-        }
-        // Integer color buffers are not emulatable via clearColor — silently no-op.
-    }
-
-    clearBufferuiv(buffer: GLenum, drawbuffer: GLint, values: Uint32List, _srcOffset?: GLuint): void {
-        const n2 = this._native2 as Gwebgl.WebGL2RenderingContext & OptionalClearBufferNative;
-        if (typeof n2.clearBufferuiv === 'function') {
-            n2.clearBufferuiv(buffer, drawbuffer, Array.from(values) as number[]);
-            return;
-        }
-        // Unsigned integer color buffers are not emulatable via clearColor —
-        // silently no-op.
-        void buffer; void drawbuffer;
-    }
-
-    clearBufferfi(buffer: GLenum, drawbuffer: GLint, depth: GLfloat, stencil: GLint): void {
-        const n2 = this._native2 as Gwebgl.WebGL2RenderingContext & OptionalClearBufferNative;
-        if (typeof n2.clearBufferfi === 'function') {
-            n2.clearBufferfi(buffer, drawbuffer, depth, stencil);
-            return;
-        }
-        // Only DEPTH_STENCIL makes sense for this entry point.
-        if (buffer === 0x84F9 /* DEPTH_STENCIL */) {
-            const prevDepth = this.getParameter(this.DEPTH_CLEAR_VALUE) as number | null;
-            const prevStencil = this.getParameter(this.STENCIL_CLEAR_VALUE) as number | null;
-            this.clearDepth(depth);
-            this.clearStencil(stencil);
-            this.clear(this.DEPTH_BUFFER_BIT | this.STENCIL_BUFFER_BIT);
-            if (prevDepth !== null) this.clearDepth(prevDepth);
-            if (prevStencil !== null) this.clearStencil(prevStencil);
-        }
-        void drawbuffer;
-    }
-
-    invalidateFramebuffer(target: GLenum, attachments: GLenum[]): void {
-        this._native2.invalidateFramebuffer(target, attachments);
-    }
-
-    invalidateSubFramebuffer(target: GLenum, attachments: GLenum[], x: GLint, y: GLint, width: GLsizei, height: GLsizei): void {
-        this._native2.invalidateSubFramebuffer(target, attachments, x, y, width, height);
-    }
-
-    readBuffer(src: GLenum): void {
-        this._native2.readBuffer(src);
-    }
-
-    renderbufferStorageMultisample(target: GLenum, samples: GLsizei, internalFormat: GLenum, width: GLsizei, height: GLsizei): void {
-        if (target !== this.RENDERBUFFER) {
-            this.setError(this.INVALID_ENUM);
-            return;
-        }
-        const renderbuffer = this._activeRenderbuffer;
-        if (!renderbuffer) {
-            this.setError(this.INVALID_OPERATION);
-            return;
-        }
-
-        this._saveError();
-        this._native2.renderbufferStorageMultisample(target, samples, internalFormat, width, height);
-        const error = this.getError();
-        this._restoreError(error);
-        if (error !== this.NO_ERROR) return;
-
-        renderbuffer._width = width;
-        renderbuffer._height = height;
-        renderbuffer._format = internalFormat;
-    }
-
-    // ─── Unsigned Integer Uniforms ────────────────────────────────────────
-
-    uniform1ui(location: WebGLUniformLocation | null, v0: GLuint): void {
-        if (!location) return;
-        this._native2.uniform1ui((location as WebGLUniformLocation)._, v0);
-    }
-
-    uniform2ui(location: WebGLUniformLocation | null, v0: GLuint, v1: GLuint): void {
-        if (!location) return;
-        this._native2.uniform2ui((location as WebGLUniformLocation)._, v0, v1);
-    }
-
-    uniform3ui(location: WebGLUniformLocation | null, v0: GLuint, v1: GLuint, v2: GLuint): void {
-        if (!location) return;
-        this._native2.uniform3ui((location as WebGLUniformLocation)._, v0, v1, v2);
-    }
-
-    uniform4ui(location: WebGLUniformLocation | null, v0: GLuint, v1: GLuint, v2: GLuint, v3: GLuint): void {
-        if (!location) return;
-        this._native2.uniform4ui((location as WebGLUniformLocation)._, v0, v1, v2, v3);
-    }
-
-    uniform1uiv(location: WebGLUniformLocation | null, data: Uint32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Uint32Array ? data : new Uint32Array(data as number[]);
-        // TODO: update @girs/gwebgl-0.1 types to accept Uint32Array; cast until then
-        this._native2.uniform1uiv((location as WebGLUniformLocation)._, arr.length, arr as unknown as number[]);
-    }
-
-    uniform2uiv(location: WebGLUniformLocation | null, data: Uint32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Uint32Array ? data : new Uint32Array(data as number[]);
-        this._native2.uniform2uiv((location as WebGLUniformLocation)._, arr.length / 2, arr as unknown as number[]);
-    }
-
-    uniform3uiv(location: WebGLUniformLocation | null, data: Uint32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Uint32Array ? data : new Uint32Array(data as number[]);
-        this._native2.uniform3uiv((location as WebGLUniformLocation)._, arr.length / 3, arr as unknown as number[]);
-    }
-
-    uniform4uiv(location: WebGLUniformLocation | null, data: Uint32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Uint32Array ? data : new Uint32Array(data as number[]);
-        this._native2.uniform4uiv((location as WebGLUniformLocation)._, arr.length / 4, arr as unknown as number[]);
-    }
-
-    // ─── Non-square Matrix Uniforms ───────────────────────────────────────
-
-    uniformMatrix2x3fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        // TODO: update @girs/gwebgl-0.1 types to accept Float32Array; cast until then
-        this._native2.uniformMatrix2x3fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
-    }
-
-    uniformMatrix3x2fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        this._native2.uniformMatrix3x2fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
-    }
-
-    uniformMatrix2x4fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        this._native2.uniformMatrix2x4fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
-    }
-
-    uniformMatrix4x2fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        this._native2.uniformMatrix4x2fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
-    }
-
-    uniformMatrix3x4fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        this._native2.uniformMatrix3x4fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
-    }
-
-    uniformMatrix4x3fv(location: WebGLUniformLocation | null, transpose: GLboolean, data: Float32List, _srcOffset?: GLuint, _srcLength?: GLuint): void {
-        if (!location) return;
-        const arr = data instanceof Float32Array ? data : new Float32Array(data as number[]);
-        this._native2.uniformMatrix4x3fv((location as WebGLUniformLocation)._, transpose, arr as unknown as number[]);
     }
 
     // ─── getUniform — WebGL2 uint type support ────────────────────────────
@@ -1242,4 +851,16 @@ export class WebGL2RenderingContext extends WebGLContextBase implements WebGL2Re
 // the published type.
 import './webgl2-context/object-lifecycle.js';
 import { installObjectLifecycleMethods } from './webgl2-context/object-lifecycle.js';
+import './webgl2-context/tex-image-3d.js';
+import { installTexImage3DMethods } from './webgl2-context/tex-image-3d.js';
+import './webgl2-context/uniform.js';
+import { installUniformMethods } from './webgl2-context/uniform.js';
+import './webgl2-context/clear-buffer.js';
+import { installClearBufferMethods } from './webgl2-context/clear-buffer.js';
+import './webgl2-context/instanced-and-framebuffer-ops.js';
+import { installInstancedAndFramebufferOpsMethods } from './webgl2-context/instanced-and-framebuffer-ops.js';
 installObjectLifecycleMethods(WebGL2RenderingContext.prototype);
+installTexImage3DMethods(WebGL2RenderingContext.prototype);
+installUniformMethods(WebGL2RenderingContext.prototype);
+installClearBufferMethods(WebGL2RenderingContext.prototype);
+installInstancedAndFramebufferOpsMethods(WebGL2RenderingContext.prototype);
